fix helpers tests
